requirements.txt
mlflow>=2.12.2
# boto3>=1.34.61 # required by mlflow, but not in AWS
onnxruntime>=1.18.0
pandas>=2.2.2
Pillow>=10.3.0
Requests>=2.32.3
streamlit>=1.35.0
torch>=2.3.1
torchvision>=0.18.1
